requirements.txt
agno==1.6.3
annotated-types==0.7.0
anyio==4.9.0
beautifulsoup4==4.13.4
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.0
curl_cffi==0.11.3
cycler==0.12.1
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
duckduckgo_search==8.0.4
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.13
fastapi-cli==0.0.7
fonttools==4.58.4
frozendict==2.4.6
gitdb==4.0.12
GitPython==3.1.44
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
jiter==0.10.0
kiwisolver==1.4.7
lxml==5.4.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
mdurl==0.1.2
multitasking==0.0.11
numpy==2.0.2
openai==1.88.0
packaging==25.0
pandas==2.3.0
peewee==3.18.1
pillow==11.2.1
platformdirs==4.3.8
primp==0.15.0
protobuf==6.31.1
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.9.1
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.7
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.46.2
tomli==2.2.1
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
yfinance==0.2.63
zipp==3.23.0
